Correction erreurs diverses

+++ inc/reader.h
@@ -15,11 +15,11 @@
 
 typedef enum READER_Status READER_Status;
 enum READER_Status{
-	READER_OK                   = (uint32_t)(0x00000001),
-	READER_NO                   = (uint32_t)(0x00000000),
-	READER_ERR                  = (uint32_t)(0x00000002),
-	READER_TIMEOUT              = (uint32_t)(0x00000003),
-	READER_BUSY                 = (uint32_t)(0x00000004)
+	READER_OK                   = (0x00000001),
+	READER_NO                   = (0x00000000),
+	READER_ERR                  = (0x00000002),
+	READER_TIMEOUT              = (0x00000003),
+	READER_BUSY                 = (0x00000004)
 };
